requirements.txt
redis>=5.0.3
python-dotenv
pymongo
falconlogger @ git+https://github.com/tjdaley/FalconLogger